@@ -1,7 +1,6 @@
 package com.hea3ven.buildingbricks.compat.vanilla;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -19,7 +18,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.common.config.Property;
 import net.minecraftforge.common.config.Property.Type;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.relauncher.Side;
@@ -67,35 +65,26 @@ public class ProxyModBuildingBricksCompatVanilla extends ProxyModModule {
 
 	@Override
 	public CategoryConfigManagerBuilder getConfig() {
-		return new CategoryConfigManagerBuilder("vanilla")
-				.addValue("replaceGrassTexture", "true", Type.BOOLEAN,
-						"Enable to replace the grass texture with a long grass texture if the " +
-								"generateGrassSlabs is enabled",
-						new Consumer<Property>() {
-							@Override
-							public void accept(final Property property) {
-								SidedCall.run(Side.CLIENT, new SidedRunnable() {
-									@Override
-									@SideOnly(Side.CLIENT)
-									public void run() {
-										LongGrassTextureGenerator.enabled = property.getBoolean();
-									}
-								});
-							}
-						})
+		return new CategoryConfigManagerBuilder("vanilla").addValue("replaceGrassTexture", "true",
+				Type.BOOLEAN, "Enable to replace the grass texture with a long grass texture if the " +
+						"generateGrassSlabs is enabled", property -> {
+					SidedCall.run(Side.CLIENT, new SidedRunnable() {
+						@Override
+						@SideOnly(Side.CLIENT)
+						public void run() {
+							LongGrassTextureGenerator.enabled = property.getBoolean();
+						}
+					});
+				})
 				.addValue("replaceGrassTextureForce", "false", Type.BOOLEAN,
-						"Enable to replace the grass texture with a long grass texture always",
-						new Consumer<Property>() {
-							@Override
-							public void accept(final Property property) {
-								SidedCall.run(Side.CLIENT, new SidedRunnable() {
-									@Override
-									@SideOnly(Side.CLIENT)
-									public void run() {
-										LongGrassTextureGenerator.forceEnabled = property.getBoolean();
-									}
-								});
-							}
+						"Enable to replace the grass texture with a long grass texture always", property -> {
+							SidedCall.run(Side.CLIENT, new SidedRunnable() {
+								@Override
+								@SideOnly(Side.CLIENT)
+								public void run() {
+									LongGrassTextureGenerator.forceEnabled = property.getBoolean();
+								}
+							});
 						})
 				.addValue("generateGrassSlabs", "true", Type.BOOLEAN,
 						"Enable to generate grass slabs in the world to smooth out the surface",
